controllers/user/me: Move `confirm_user_email()` to dedicated module
