+++ src/controllers/user.rs
@@ -1,9 +1,9 @@
 pub mod email_notifications;
+pub mod email_verification;
 pub mod me;
 pub mod other;
-pub mod resend;
 pub mod session;
 pub mod update;
 
-pub use resend::resend_email_verification;
+pub use email_verification::resend_email_verification;
 pub use update::update_user;
